chat textField, jbutton, jlabel
Added chat interface.
text field to enter message
button to send message
and jlabel to view message history.
currently label overlaps the bottom left corner. I am hoping to figure
out a way to make it semiopque when using the text field. worst case we
would need to make the Frame larger as there isnt much room for the chat
feature. or make it in a different window.