started pres
pretty similar to before.
Added a few introduction to chess slides.
I think it looks pretty good.
Please add your slides as you are ready.

Also I tried to change the fonts on the titles but it seems that its not
as trivial as changing it to font.Motorwerk for example and i didnt want
to risk using a font that is not installed on the school computer. So I
left that.